requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
numpy==1.24.3
pandas==2.1.3
python-dotenv==1.0.0
requests==2.31.0
nltk==3.8.1
scikit-learn==1.3.2
streamlit==1.28.1
openai==1.3.5
